fix: restructure test namespaces to match Outcomes API changes
- Moved test files to match new API structure:
  - tests/Api/OutcomeGroups/ for OutcomeGroup tests
  - tests/Api/OutcomeImports/ for OutcomeImport tests
  - tests/Api/OutcomeResults/ for OutcomeResult tests
  - tests/Dto/OutcomeGroups/ for OutcomeGroup DTO tests
- Updated all test namespaces to match new directory structure
- All tests pass with no PHPStan errors
