+++ tests/Api/OutcomeImports/OutcomeImportTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CanvasLMS\Tests\Api\Outcomes;
+namespace CanvasLMS\Tests\Api\OutcomeImports;
 
 use CanvasLMS\Api\OutcomeImports\OutcomeImport;
 use CanvasLMS\Config;
