+++ tests/Dto/Outcomes/UpdateOutcomeDTOTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CanvasLMS\Tests\Dto\Outcomes\Outcome;
+namespace CanvasLMS\Tests\Dto\Outcomes;
 
 use CanvasLMS\Dto\Outcomes\UpdateOutcomeDTO;
 use PHPUnit\Framework\TestCase;
